refactor: Lines 컴포넌트에서 LinesList 컴포넌트 분리
